requirements.txt
streamlit
requests
llama-index
matplotlib
scikit-image
pillow
langchain-community